oppdater navn på mappe for oppgaveliste
