+++ apps/etterlatte-saksbehandling-ui/client/src/components/oppgavebenk/FilterRad.tsx
@@ -15,7 +15,7 @@ import {
   SAKSBEHANDLERFILTER,
   YTELSEFILTER,
   YtelseFilterKeys,
-} from '~components/nyoppgavebenk/Oppgavelistafiltre'
+} from '~components/oppgavebenk/Oppgavelistafiltre'
 import { useFeatureEnabledMedDefault } from '~shared/hooks/useFeatureToggle'
 import { FEATURE_TOGGLE_KAN_BRUKE_KLAGE } from '~components/person/OpprettKlage'
 import { FlexRow } from '~shared/styled'
