+++ apps/etterlatte-saksbehandling-ui/client/src/components/oppgavebenk/tildeling/ByttSaksbehandler.tsx
@@ -1,11 +1,12 @@
+import { useAppSelector } from '~store/Store'
 import { useApiCall } from '~shared/hooks/useApiCall'
-import { fjernSaksbehandlerApi } from '~shared/api/oppgaver'
+import { byttSaksbehandlerApi } from '~shared/api/oppgaver'
 import { Alert, Button, Label, Loader } from '@navikt/ds-react'
 import { PencilIcon } from '@navikt/aksel-icons'
 import React, { useEffect, useState } from 'react'
 import { GeneriskModal } from '~shared/modal/modal'
 import styled from 'styled-components'
-import { RedigerSaksbehandlerProps } from '~components/nyoppgavebenk/tildeling/RedigerSaksbehandler'
+import { RedigerSaksbehandlerProps } from '~components/oppgavebenk/tildeling/RedigerSaksbehandler'
 
 import { isSuccess, mapAllApiResult } from '~shared/api/apiUtils'
 import { ApiErrorAlert } from '~ErrorBoundary'
@@ -15,22 +16,23 @@ const SaksbehandlerWrapper = styled(Label)`
   margin-right: 0.5rem;
 `
 
-export const FjernSaksbehandler = (props: RedigerSaksbehandlerProps) => {
+export const ByttSaksbehandler = (props: RedigerSaksbehandlerProps) => {
   const [modalIsOpen, setModalIsOpen] = useState(false)
-  const { saksbehandler, oppgaveId, sakId, oppdaterTildeling, erRedigerbar, versjon, type } = props
-  const [fjernSaksbehandlerSvar, fjernSaksbehandler] = useApiCall(fjernSaksbehandlerApi)
+  const { saksbehandler, oppgaveId, oppdaterTildeling, erRedigerbar, versjon, type } = props
+  const innloggetSaksbehandler = useAppSelector((state) => state.saksbehandlerReducer.innloggetSaksbehandler)
+  const [byttSaksbehandlerSvar, byttSaksbehandler] = useApiCall(byttSaksbehandlerApi)
 
   useEffect(() => {
-    if (isSuccess(fjernSaksbehandlerSvar)) {
-      oppdaterTildeling(oppgaveId, null, fjernSaksbehandlerSvar.data.versjon)
+    if (isSuccess(byttSaksbehandlerSvar)) {
+      oppdaterTildeling(oppgaveId, innloggetSaksbehandler.ident, byttSaksbehandlerSvar.data.versjon)
     }
-  }, [fjernSaksbehandlerSvar])
+  }, [byttSaksbehandlerSvar])
 
   return (
     <>
       {mapAllApiResult(
-        fjernSaksbehandlerSvar,
-        <Loader size="small" title="Fjerner saksbehandler" />,
+        byttSaksbehandlerSvar,
+        <Loader size="small" title="Bytter saksbehandler" />,
         <>
           {erRedigerbar ? (
             <Button
@@ -48,20 +50,26 @@ export const FjernSaksbehandler = (props: RedigerSaksbehandlerProps) => {
 
           <GeneriskModal
             tittel="Endre saksbehandler"
-            beskrivelse="Ønsker du å fjerne deg som saksbehandler?"
-            tekstKnappJa="Ja, fjern meg som saksbehandler"
+            beskrivelse={`Vil du overta oppgaven til ${saksbehandler}?`}
+            tekstKnappJa="Ja, overta oppgaven"
             tekstKnappNei="Nei"
-            onYesClick={() => fjernSaksbehandler({ oppgaveId: oppgaveId, sakId: sakId, type: type, versjon: versjon })}
+            onYesClick={() =>
+              byttSaksbehandler({
+                oppgaveId: oppgaveId,
+                type: type,
+                nysaksbehandler: { saksbehandler: innloggetSaksbehandler.ident, versjon: versjon },
+              })
+            }
             setModalisOpen={setModalIsOpen}
             open={modalIsOpen}
           />
         </>,
         () => (
-          <ApiErrorAlert size="small">Feil ved fjerning av saksbehandling</ApiErrorAlert>
+          <ApiErrorAlert size="small">Kunne ikke bytte saksbehandler fra oppgaven</ApiErrorAlert>
         ),
         () => (
           <Alert variant="success" size="small">
-            Du er fjernet som saksbehandler
+            Saksbehandler er endret og oppgaven ble lagt på din oppgaveliste
           </Alert>
         )
       )}
