+++ apps/etterlatte-saksbehandling-ui/client/src/components/oppgavebenk/tildeling/FjernSaksbehandler.tsx
@@ -1,12 +1,11 @@
-import { useAppSelector } from '~store/Store'
 import { useApiCall } from '~shared/hooks/useApiCall'
-import { byttSaksbehandlerApi } from '~shared/api/oppgaver'
+import { fjernSaksbehandlerApi } from '~shared/api/oppgaver'
 import { Alert, Button, Label, Loader } from '@navikt/ds-react'
 import { PencilIcon } from '@navikt/aksel-icons'
 import React, { useEffect, useState } from 'react'
 import { GeneriskModal } from '~shared/modal/modal'
 import styled from 'styled-components'
-import { RedigerSaksbehandlerProps } from '~components/nyoppgavebenk/tildeling/RedigerSaksbehandler'
+import { RedigerSaksbehandlerProps } from '~components/oppgavebenk/tildeling/RedigerSaksbehandler'
 
 import { isSuccess, mapAllApiResult } from '~shared/api/apiUtils'
 import { ApiErrorAlert } from '~ErrorBoundary'
@@ -16,23 +15,22 @@ const SaksbehandlerWrapper = styled(Label)`
   margin-right: 0.5rem;
 `
 
-export const ByttSaksbehandler = (props: RedigerSaksbehandlerProps) => {
+export const FjernSaksbehandler = (props: RedigerSaksbehandlerProps) => {
   const [modalIsOpen, setModalIsOpen] = useState(false)
-  const { saksbehandler, oppgaveId, oppdaterTildeling, erRedigerbar, versjon, type } = props
-  const innloggetSaksbehandler = useAppSelector((state) => state.saksbehandlerReducer.innloggetSaksbehandler)
-  const [byttSaksbehandlerSvar, byttSaksbehandler] = useApiCall(byttSaksbehandlerApi)
+  const { saksbehandler, oppgaveId, sakId, oppdaterTildeling, erRedigerbar, versjon, type } = props
+  const [fjernSaksbehandlerSvar, fjernSaksbehandler] = useApiCall(fjernSaksbehandlerApi)
 
   useEffect(() => {
-    if (isSuccess(byttSaksbehandlerSvar)) {
-      oppdaterTildeling(oppgaveId, innloggetSaksbehandler.ident, byttSaksbehandlerSvar.data.versjon)
+    if (isSuccess(fjernSaksbehandlerSvar)) {
+      oppdaterTildeling(oppgaveId, null, fjernSaksbehandlerSvar.data.versjon)
     }
-  }, [byttSaksbehandlerSvar])
+  }, [fjernSaksbehandlerSvar])
 
   return (
     <>
       {mapAllApiResult(
-        byttSaksbehandlerSvar,
-        <Loader size="small" title="Bytter saksbehandler" />,
+        fjernSaksbehandlerSvar,
+        <Loader size="small" title="Fjerner saksbehandler" />,
         <>
           {erRedigerbar ? (
             <Button
@@ -50,26 +48,20 @@ export const ByttSaksbehandler = (props: RedigerSaksbehandlerProps) => {
 
           <GeneriskModal
             tittel="Endre saksbehandler"
-            beskrivelse={`Vil du overta oppgaven til ${saksbehandler}?`}
-            tekstKnappJa="Ja, overta oppgaven"
+            beskrivelse="Ønsker du å fjerne deg som saksbehandler?"
+            tekstKnappJa="Ja, fjern meg som saksbehandler"
             tekstKnappNei="Nei"
-            onYesClick={() =>
-              byttSaksbehandler({
-                oppgaveId: oppgaveId,
-                type: type,
-                nysaksbehandler: { saksbehandler: innloggetSaksbehandler.ident, versjon: versjon },
-              })
-            }
+            onYesClick={() => fjernSaksbehandler({ oppgaveId: oppgaveId, sakId: sakId, type: type, versjon: versjon })}
             setModalisOpen={setModalIsOpen}
             open={modalIsOpen}
           />
         </>,
         () => (
-          <ApiErrorAlert size="small">Kunne ikke bytte saksbehandler fra oppgaven</ApiErrorAlert>
+          <ApiErrorAlert size="small">Feil ved fjerning av saksbehandling</ApiErrorAlert>
         ),
         () => (
           <Alert variant="success" size="small">
-            Saksbehandler er endret og oppgaven ble lagt på din oppgaveliste
+            Du er fjernet som saksbehandler
           </Alert>
         )
       )}
